Move services to Logic project
