+++ Backend/Api/Controllers/AdminController.cs
@@ -1,4 +1,4 @@
-﻿using Api.Services.Implementations;
+﻿using Logic.Services.Implementations;
 using Domain.Admin;
 using Microsoft.AspNetCore.Mvc;
