+++ Backend/Api/Controllers/ProviderController.cs
@@ -1,4 +1,4 @@
-﻿using Api.Services.Implementations;
+﻿using Logic.Services.Implementations;
 using Domain.Common;
 using Microsoft.AspNetCore.Mvc;
