+++ Backend/Api/Controllers/TicketController.cs
@@ -1,4 +1,4 @@
-﻿using Api.Services.Implementations;
+﻿using Logic.Services.Implementations;
 using Domain.User;
 using Microsoft.AspNetCore.Mvc;
