--- Backend/Api/Services/Implementations/UserService.cs
@@ -1,28 +1,0 @@
-﻿using Domain.User;
-
-namespace Api.Services.Implementations
-{
-    public class UserService
-    {
-        public bool CreateUserAccount(User user)
-        {
-            throw new NotImplementedException();
-        }
-        public bool RemoveUserAccount(int userID)
-        {
-            throw new NotImplementedException();
-        }
-        public bool VerifyUserAccount(int userID)
-        {
-            throw new NotImplementedException();
-        }
-        public bool AuthoriseUser(int userID)
-        {
-            throw new NotImplementedException();
-        }
-        public User GetUserByID(int userID)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
