+++ Backend/Logic/Services/Implementations/AdminService.cs
@@ -1,0 +1,28 @@
+﻿using Domain.Admin;
+
+namespace Logic.Services.Implementations
+{
+    public class AdminService
+    {
+        public bool CreateAdminAccount(Admin admin)
+        {
+            throw new NotImplementedException();
+        }
+        public bool RemoveAdminAccount(int adminID)
+        {
+            throw new NotImplementedException();
+        }
+        public bool VerifyAdminAccount(int adminID)
+        {
+            throw new NotImplementedException();
+        }
+        public bool AuthoriseAdmin(int adminID)
+        {
+            throw new NotImplementedException();
+        }
+        public void GiveAdminPermissions(Admin admin)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
