+++ Backend/Logic/Services/Implementations/ProviderService.cs
@@ -1,6 +1,6 @@
 ﻿using Domain.Common;
 
-namespace Api.Services.Implementations
+namespace Logic.Services.Implementations
 {
     public class ProviderService
     {
