+++ Backend/Logic/Services/Implementations/TicketService.cs
@@ -1,6 +1,6 @@
 ﻿using Domain.User;
 
-namespace Api.Services.Implementations
+namespace Logic.Services.Implementations
 {
     public class TicketService
     {
